+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Add metrics & logs relevant for RPCh debugging ([#4995](https://github.com/hoprnet/hoprnet/pull/4995))
 - Enhance debug logs in dialing logic to enhance debugging ([#5004](https://github.com/hoprnet/hoprnet/pull/5004))
 - Fix address coercion issue in libp2p address handling ([#5020](https://github.com/hoprnet/hoprnet/pull/5020))
+- Add send message over websocket support ([#4882](https://github.com/hoprnet/hoprnet/pull/4882))
 
 <a name="1.92"></a>